+++ apps/pie-storybook/stories/testing/pie-radio-group.test.stories.ts
@@ -10,6 +10,7 @@ import '@justeattakeaway/pie-link';
 import '@justeattakeaway/pie-radio';
 import '@justeattakeaway/pie-form-label';
 import '@justeattakeaway/pie-button';
+import '@justeattakeaway/pie-icons-webc/dist/IconPlusCircle';
 
 import { createStory } from '../../utilities';
 
@@ -48,7 +49,7 @@ export default radioGroupStoryMeta;
 const KeyboardNavigationTemplate = () => html`
     <h2>Radio group 1</h2>
     <p><pie-button size="small-productive" data-test-id="btn-1">Button 1</pie-button></p>
-    <pie-radio-group data-test-id="radio-group-1">
+    <pie-radio-group name="radio-group-1" data-test-id="radio-group-1">
         <pie-radio data-test-id="radio-1" value="chinese">Chinese</pie-radio>
         <pie-radio data-test-id="radio-2" value="shawarma">Shawarma</pie-radio>
         <pie-radio data-test-id="radio-3" value="pizza">Pizza</pie-radio>
@@ -69,4 +70,38 @@ const KeyboardNavigationTemplate = () => html`
     <p><pie-button size="small-productive" variant="secondary" data-test-id="btn-4">Button 4</pie-button></p>
     `;
 
+const DynamicSlotsTemplate = () => {
+    let counter = 2;
+
+    // Function to dynamically add radio buttons to the radio group
+    const onAddButtonClick = () => {
+        const radioGroup = document.querySelector('pie-radio-group[data-test-id="radio-group-1"]');
+
+        if (radioGroup) {
+            const newRadio = document.createElement('pie-radio');
+            newRadio.setAttribute('value', counter.toString());
+            newRadio.textContent = `Option ${counter}`;
+
+            radioGroup.appendChild(newRadio);
+
+            counter++;
+        }
+    };
+
+    return html`
+            <pie-radio-group name="radio-group-1" data-test-id="radio-group-1">
+                <pie-radio data-test-id="radio-1" value="1">Option 1</pie-radio>
+            </pie-radio-group>
+
+            <p>
+                <pie-button data-test-id="add-option" iconplacement="leading" @click="${onAddButtonClick}">
+                    <icon-plus-circle slot="icon"></icon-plus-circle>
+                    Add another option
+                </pie-button>
+            </p>
+        `;
+};
+
 export const KeyboardNavigation = createStory<RadioGroupProps>(KeyboardNavigationTemplate, defaultArgs)();
+export const DynamicSlots = createStory<RadioGroupProps>(DynamicSlotsTemplate, defaultArgs)();
+